validation-results/medium_rest_api_5e77b544_20251105_141713/code/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.7.0
bcrypt==4.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
